change way targetpage and targetlist get data
