+++ src/App.jsx
@@ -1,12 +1,11 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import {
   Switch,
   Route,
   Redirect,
-  useRouteMatch,
 } from 'react-router-dom';
 import Header from './components/Navigation';
-import TargetPage from './components/TargetPage';
+import Target from './components/TargetPage/index';
 import Home from './components/Home';
 import targetService from './services/targets';
 import NewTargetForm from './components/NewTargetForm';
@@ -15,26 +14,9 @@ import './assets/styles/App.css';
 import AdminPanel from './components/AdminPanel/AdminPanel';
 
 function App() {
-  const [targets, setTargets] = useState('loading...');
-
-  const getTargets = async () => {
-    const data = await targetService.getAllTargets();
-    data.features.sort((a, b) => (a.properties.name > b.properties.name ? 1 : -1));
-    setTargets(data.features);
-  };
-
   const createNewTarget = (newTarget) => {
     targetService.postTarget(newTarget);
   };
-
-  useEffect(() => {
-    getTargets();
-  }, []);
-
-  const match = useRouteMatch('/hylyt/:id');
-  const target = match && targets !== 'loading...'
-    ? targets.find((t) => t.properties.id === match.params.id)
-    : null;
 
   return (
     <div className="container">
@@ -44,11 +26,11 @@ function App() {
           <Redirect to="/etusivu" />
         </Route>
         <Route path="/etusivu" component={Home} />
-        <Route path="/hylyt/:id">
-          <TargetPage target={target} />
+        <Route path="/hylyt/:id" component={Target}>
+          {/* <TargetPage /> */}
         </Route>
         <Route path="/hylyt">
-          <TargetList targets={targets} />
+          <TargetList />
         </Route>
         <Route path="/admin">
           <AdminPanel />
